Fall back to an 'other' plural lookup before giving up
This change alters the v1 behavior so that plural lookups where
the plural count is greater than 1, but no translation is found,
will attempt another translation with 'other' before falling back
to the default language. This will only happen if the first
translation attempt was not 'other' to begin with.