Add a nil pointer check before dereferencing an identifier's Obj
When walking the AST during extraction, if an identifier in a composite
literal is encountered that does not define in that file, the `Obj`
field of the `ast.Ident` will be nil.

Rather than panicking in this case, have the literal string extraction
function return false (no string found).

Also, add a test case representing this scenario.