Error if reserved message keys are mixed with unreserved keys (#320)
Makes it an error to provide data that mixes reserved message keys with
unreserved ones. Previously the unreserved keys would be ignored, but
this caused surprising behavior where if someone added a message that
had a key with a reserved word, it would silently break the whole
translation file. This change turns that situation into an error.

Resolves https://github.com/nicksnyder/go-i18n/issues/209